unify series cards
