+++ packages/browser/src/components/series/index.ts
@@ -1,2 +1,4 @@
+export { default as LibrarySeriesCard } from './LibrarySeriesCard'
 export { default as SeriesBooksAlphabet } from './SeriesBooksAlphabet'
+export { StackedSeriesCard } from './StackedSeriesCard'
 export { SeriesTable } from './table'
